backend/pyproject.toml
[project]
name = "url-crawl-embed"
version = "0.1.0"
description = "Backend pipeline to crawl Docusaurus book URLs, extract text, generate embeddings, and store in Qdrant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]